migrated file to ts
